Nexa_Deep_Research_Agent/pyproject.toml
[project]
name = "nexa-deep-research-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "redis==5.0.1",
    "aioredis==2.0.1",
    "asyncio-background-tasks==1.0.0",
    "httpx==0.25.2",
    "openai==1.3.7",
    "stripe==7.8.0",
    "typer==0.9.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "sentence-transformers==2.2.2",
    "psycopg2-binary==2.9.9",
    "sqlalchemy==2.0.23",
    "alembic==1.13.1"
]
